Create option to use legacy sts mock version of script or new test-resources
